backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
motor==3.3.2
pymongo==4.6.2
beanie==1.25.0
pydantic==2.6.4
pydantic-settings==2.2.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
aiofiles==23.2.1
Pillow==10.3.0
pytesseract==0.3.10
easyocr==1.7.1
scikit-learn==1.4.2
pandas==2.2.2
numpy==1.26.4
prophet==1.1.5
python-dotenv==1.0.1
httpx==0.27.0
pytest==8.1.1
pytest-asyncio==0.23.6
pdf2image==1.17.0
reportlab==4.1.0
openpyxl==3.1.2
